Copy core utilities for browser build
