+++ src/core/state.js
@@ -1,0 +1,44 @@
+/**
+ * Check if a value is a non-null object.
+ * @param {*} value - Value to test.
+ * @returns {boolean} True when the value is a non-null object.
+ */
+export function isNonNullObject(value) {
+  return Boolean(value) && typeof value === 'object';
+}
+
+function bothAreNotArrays(a, b) {
+  return !Array.isArray(a) && !Array.isArray(b);
+}
+
+function bothAreNonNullObjects(a, b) {
+  return isNonNullObject(a) && isNonNullObject(b);
+}
+
+function shouldDeepMerge(targetValue, sourceValue) {
+  return (
+    bothAreNonNullObjects(targetValue, sourceValue) &&
+    bothAreNotArrays(targetValue, sourceValue)
+  );
+}
+
+/**
+ * Deeply merges two objects, producing a new object.
+ * @param {object} target - Destination object.
+ * @param {object} source - Source object to merge.
+ * @returns {object} The merged object.
+ */
+export function deepMerge(target, source) {
+  const output = { ...target };
+  const mergeKey = key => {
+    const targetValue = target[key];
+    const sourceValue = source[key];
+    if (shouldDeepMerge(targetValue, sourceValue)) {
+      output[key] = deepMerge(targetValue, sourceValue);
+    } else {
+      output[key] = sourceValue;
+    }
+  };
+  Object.keys(source).forEach(mergeKey);
+  return output;
+}
